Refactor into separate modules.
